requirements.txt
mysql-connector-python
streamlit
dotenv